try fix files processing

+++ src/main/java/com/educationapp/server/models/api/FileApi.java
@@ -18,7 +18,7 @@ public class FileApi {
     private String fileName;
 
     @NotNull
-    private Long type;
+    private Integer type;
 
     @NotNull
     private Long subjectId;
